Don't change model.group on group change
Don't change model.group in AnnotationController when the currently focused
group changes.

model is supposed to be a read-only domain model, change vm.annotation instead.

Changes to vm.annotation are copied over to model before saving model to the
server.

To make this work vm.group() should return the current group of the annotation,
accounting for any changes made to vm.annotation but not saved to model or to
the server yet. It should return vm.annotation.group not model.group.

Similarly vm.isShared() should be based on vm.annotation not on model.

Also remove some code in thread.coffee that was deliberately hiding new
annotations that don't belong to the focused group (see commit
272b2432c48a2d6bf890daa863e92f3a678b27b5). This is no longer the desired
behaviour - when changing groups any new annotations move to the new group with
us (and the GROUP_FOCUSED listener in AnnotationController will update
vm.annotation.group to match).