Don't change domainModel.group on group focused
Don't change the value of domainModel.group when the focused group changes.

Instead, change it to groups.focusedGroup() in updatedDomainModel() before the
annotation is sent to the server.

This fits with the idea that domainModel is a read-only copy of the annotation
that we received from the server, and vm is the read-write copy. But there's no
vm.group as this would always be the same as groups.focusedGroup().