Move groups annotation transformers into `h.api.transform`
The `h.api.groups.logic` package provided a set of functions to
transform incoming annotation data. Most such functions already live in
`h.api.transform`, and this commit consolidates these functions into one
place.

The principal motivation for doing this is to enable the primary client
of the `h.api.transform` module, `h.api.storage`, to pass a "fetcher"
function to the one transformer that requires storage access, rather
than having a circular dependency in which `h.api.storage` require
transformers, which in turn require storage access.

As a side-effect of this move, it seems silly to keep the
`group_principals` function buried in `h.api.groups.auth` when in
reality it depends solely on the `.groups` property of the user model.
As such, this has moved to `h.auth`.