Send activation emails from a worker process
Send new user account activation emails from a worker process, instead of from
the main application thread.

This avoids a problem when the register view is called and sending the email
(mailer.send(message)) raises an exception *but the email was actually sent*.
The transaction would be rolled back and the user account not created, and the
user would receive an unusable "please activate your account" email.

Changes:

- The register view callable now puts a message on a new "activations"
  nsq topic, instead of actually sending the email.

- The new h/accounts/worker.py is a hypothesis-worker that subscribes to the
  new "activations" topic and actually sends an email for each message
  published to the toipic.

- Update accounts/test/views_test.py to fix broken tests.
  These were mainly broken because _register() now accesses
  request.get_queue_writer() and Pyramid's DummyRequest() objects don't have
  that method. Just use a mock.Mock() instead.
  In a couple of places assertions had to be updated as well.

In sending the email fails (if pyramid_mailer's send_immediately() raises) then
the message will be requeued (this is gnsq's default behavior when a worker
function raises an exception) and the worker will try again to send it.

Note that we don't attempt to prevent sending the same activation email
multiple times if a message published to the "activations" topic is received by
the worker multiple times. This may happen, since nsq guarantees
"at least once" delivery. If this becomes a problem we may need to set a flag
in the database to record when a particular activation email has already been
sent successfully.