Create standalone websocket server application
Make it possible to run the websocket server on its own, separate from
the rest of the application code.

This adds a paste entrypoint called "websocket" and a pair of config
files (for production and development) at

    conf/websocket.ini
    conf/websocket-development.ini

These allow you to run a standalone websocket server (listening by
default on port 5001 in development mode and 8001 in production mode)
using, for example:

    gunicorn --paste conf/websocket.ini

The gevent/WSGI hybrid worker class is copied from h.server, which will
be removed imminently.