Switch main web worker to default worker class
This commit removes the websocket endpoint (/ws) from the main web
application, and switches that application to gunicorn's default (sync)
worker class.

This fixes a class of problems where under heavy load, many concurrent
greenlets would exhaust the connection pools to PostgreSQL and/or
Elasticsearch.

By default, the web worker will now run in a configuration in which the
client code will not attempt to connect to the websocket. To send the
websocket URL to the client, you must set the H_WEBSOCKET_URL
environment variable. For example:

    H_WEBSOCKET_URL=wss://hypothes.is/ws gunicorn --paste conf/production.ini

This also makes the use of psycogreen and the gevented Sentry worker
conditional, so that the websocket server makes use of these while the
main web application uses the more conventional defaults.